refactor: 🔧 Jules05: Resolve cross-agent conflict in Risk Management and Feature Engineering
- Harmonized RiskManager and AuditedRiskManager with 8-layer cascade.
- Relocated FeatureEngineer to src/data/ per domain separation.
- Updated main.py to use the new RiskDecision API.
- Cleaned up redundant files and exports.
- Documented findings in CONFLICT_LOG.md.
