+++ scripts/verify_backtest_audit.py
@@ -9,7 +9,7 @@ import pandas as pd
 sys.path.append(os.getcwd())
 
 from src.core.audit_log import AuditLogger
-from src.core.feature_engineering import FeatureEngineer
+from src.data.feature_engineering import FeatureEngineer
 from src.trading.backtester import BacktestEngine
